protos: 158 -> 4 block-scope declarations across eight files
Every edit was A/B'd with objdump against the pre-edit object and every one
came back byte-identical, so no hoist changed codegen anywhere. ROM sha1
6cea2d46b929 unchanged.

The structural finding is worth more than the count. Two files kept their
'file-scope' prototypes INSIDE '#ifdef PORT' arms, so the N64 build saw none
of them. In ovl2_2.c the PORT region spans lines 29-324 and the prototype
group sat at 112-119, which left the N64 build calling func_800A9AA8,
func_800F8560, func_8010137C, func_800FA414 and func_800FBF18 IMPLICITLY.
plydemo.c had the same trap with seventeen prototypes. Both groups now sit
after the '#endif'.

Conflicts reconciled against their definitions include func_800A9AA8
(u32 (s32,s32) -> void *(u32,s32), which also removed a latent PC-side
pointer truncation), func_800A9864 (-> void (u32,u32,u32)),
utilGetTransformSRT and func_800A4DB8 (Vector * -> DObj *), func_800B5064
(void (s32) -> void (GObj *)), assign_new_process_entry
(s32 -> struct GObjProcess *, with gEntityGObjProcessArray retyped to match),
func_800FB914 (void -> s32), func_8011E244 (s32 -> u8) and func_80104C24
(s32 -> u8).

Two declarations are load-bearing and reverted with measurements in the
source: func_800AFBB4 as (u32, GObj *) -- with the definition's s32, IDO
re-materialises 'li $a0,1' in func_801DE030_ovl16 instead of reusing $s3, and
that one instruction was the entire object diff -- and func_801A3E80_ovl7 as
'(void)', because func_801BDBFC_ovl7 calls it with no arguments and an ANSI
prototype does not compile.

Four remain in ovl2_2.c, each commented: their MIPS_TO_C drafts call the
symbol with a leftover register as an extra argument, so the draft's arity
disagrees with the tree's definition. Hoisting the correct spelling would
make the TU contradict itself under MIPS_TO_C and break an active factory
seed; changing the draft's call would change what the permuter measures.
Each comment names the real definition and says to resolve the arity from
the listing first.

+++ src/ovl7/ovl7_16.c
@@ -63,9 +63,13 @@ void func_800B19F4(s32, s32);
 void ohSleep(s32);
 void func_8019D958_ovl7(u16);
 void func_800AFA14(void);
+/* `void (void)` is load-bearing here: the definition in ovl7_4.c takes
+ * GObj *, but the call in func_801BDBFC_ovl7 below passes 0 args -- the ROM
+ * relies on whatever GObj * is already in $a0, same as func_801BE490_ovl7
+ * above. An ANSI prototype breaks compilation with "too few arguments". */
 void func_801A3E80_ovl7(void);
 void func_800AFBB4(s32, struct GObj *);
-s32 func_80104C24(Vector *, Vector *);
+u8 func_80104C24(Vector *, Vector *);
 u32 func_800B9DF8(s32);
 void func_801BDEB0_ovl7(GObj *);
 extern FUNCLIST D_801CD8C0_ovl7;
